Moved metadata related code to their own file.
