+++ ingest/exporter/bundle_update_service.py
@@ -1,37 +1,7 @@
 from copy import deepcopy
-from typing import Iterable
 
 from ingest.api.dssapi import DssApi
-
-
-class MetadataResource:
-
-    def __init__(self, metadata_type=None, metadata_json=None, uuid=None, dcp_version=None):
-        self.metadata_json = metadata_json
-        self.uuid = uuid
-        self.dcp_version = dcp_version
-        self.metadata_type = metadata_type
-        if not metadata_type:
-            self._determine_metadata_type()
-
-    def _determine_metadata_type(self):
-        metadata_type = None
-        if self.metadata_json:
-            described_by = self.metadata_json.get('describedBy')
-            metadata_type = described_by.split('/')[-1] if described_by else None
-        self.metadata_type = metadata_type
-
-    @staticmethod
-    def from_dict(data: dict):
-        uuid_object = data.get('uuid')
-        uuid = uuid_object.get('uuid') if uuid_object else None
-        content = data.get('content')
-        metadata_resource = MetadataResource(uuid=uuid, metadata_json=content,
-                                             dcp_version=data.get('dcpVersion'))
-        return metadata_resource
-
-    def get_staging_file_name(self):
-        return f'{self.uuid}.{self.dcp_version}.json'
+from ingest.exporter.metadata import MetadataResource, MetadataService
 
 
 class Bundle:
@@ -67,16 +37,6 @@ class Bundle:
         target_file = self.get_file(metadata_resource.uuid)
         target_file['version'] = metadata_resource.dcp_version
         target_file['content-type'] = f'metadata/{metadata_resource.metadata_type}'
-
-
-class MetadataService:
-
-    def __init__(self, ingest_client):
-        self.ingest_client = ingest_client
-
-    def fetch_resource(self, resource_link: str) -> MetadataResource:
-        raw_metadata = self.ingest_client.get_entity_by_callback_link(resource_link)
-        return MetadataResource.from_dict(raw_metadata)
 
 
 class StagingInfo:
